Extract and clean pileup logic into testable state
There looks to be a bug where in the implementation of overlap double-counting avoidance in the pileup function.
When a read is encountered, the qname and base are recorded before the data is added to the counts array.
If the other member of the read pair also overlaps the position of interest, the data is only added to the
counts array iff the base call is different than the one seen on the first member encountered. Fine.
However, subsequent quality/position filters operating after this guard may translate a given base call to an
ambiguous N before adding the data to the counts array. This is not reflected in the overlap record,
so if the second member encountered of the pair has a better call of the same base, the data of the
second member will simply be discarded.

The first step to figuring this out is to make this unit-testable - which is the purpose of the commit.
Basically a refactor while preserving all original functionality to expose testable units and make
things a bit clearer. The bug has NOT been fixed yet. I've added a makefile and a tiny test suite to
confirm this all compiles and that the overlap guard works in the basic sense.
I've also done some brush cutting of extraneous variables and so on that do nothing (insofar as I can tell).